chang to float.32

+++ src/model/models.py
@@ -13,7 +13,7 @@ class GeneralModel:
         self.checkpoint = checkpoint
         self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
         self.tokenizer = AutoTokenizer.from_pretrained(checkpoint)
-        self.base_model = AutoModelForSeq2SeqLM.from_pretrained(checkpoint, torch_dtype=torch.float16).to(self.device)
+        self.base_model = AutoModelForSeq2SeqLM.from_pretrained(checkpoint, torch_dtype=torch.float32).to(self.device)
 
     def generate(self, input_text, **kwargs):
         try:
